+++ sys/src/cmd/cifs/main.c
@@ -1148,7 +1148,7 @@ void
 main(int argc, char **argv)
 {
 	int i, n;
-	long svrtime;
+	int32_t svrtime;
 	char windom[64], cname[64];
 	char *method, *sysname, *keyp, *mtpt, *svs;
 	static char *sh[1024];
